// src/app/templates/shared/models/designer-item-type.ts
export enum DesignerItemType {
  Static = 'static',
  Checkbox = 't_Checkbox',
  IntDropdown = 't_IntDropdown',
  StringDropdown = 't_StringDropdown',
  FieldGroup = 't_FieldGroup',
  ColumnGroup = 't_ColumnGroup',
  MultipleFieldGroup = 't_MultipleFieldGroup',
  IntText = 't_IntText',
  FloatText = 't_FloatText',
  StringText = 't_StringText',
  DateText  = 't_Date',
  Attachment = 'attachment',
}